add quiet_vehicles, reducing sound levels in vehicles
To configure, execute in debug console:

    profileNamespace setVariable ["quiet_vehicles_coef", 0.4];

where 0.4 is the coefficient (percentage) of how what to reduce
the level to compared to your Sound volume setting in the Settings
menu, from 0 (no sound) to 1 (level equal your Settings).

For example 0.4 coef with 50% Sound volume in Settings results in
20% overall Sound level.

Default coef is 1.0.

Signed-off-by: freghar <freghar@dummy.tld>