ares_extras: rewrite marker add/set/del as a simple move
Even without swt_markers, it seems that any markers present
in the mission are not moved/removed when a client JIPs into
a mission, leaving ie. an old respawn marker in place, despite
the server having removed it.

This issue is not present on a simple position change of
an existing marker.

Signed-off-by: freghar <freghar@dummy.tld>