lighting_fixes: remove flashlight customizations
Vanilla flashlight, with or without custom config changes,
has locally displayed light intensity bugs when used by other
players or AI - the first ~3 seconds are super intensive.

This is a mild issue with vanilla flashlight config, but a major
one with custom one as the high intensity lights up a good 200-300
meters ahead with intense light.

Signed-off-by: freghar <freghar@dummy.tld>