replace isEqualTo with == when comparing objects
While isEqualTo is a more low level version which is approx. 10-20%
faster, comparing objects is very lightweight even for == or != ,
so use the more legible version instead.

isEqualTo still makes sense when comparing potentially unknown data
types (check if a variable is ie. "" where it could be even [])
and when comparing arrays (where it does pointer/reference check).

Signed-off-by: freghar <freghar@dummy.tld>