ai_dynamic_skill: revert UI sliders back to vanilla 0.2-1.0 range
The range modification was done mainly for debugging purposes and while
having access to 0 skill value may make some sense sometimes (ie. unit
that never hits with bullets, even on 1m distance), it probably doesn't
make sense to have it UI-accessible as it can lead to Zeus mistakenly
setting it.

Signed-off-by: freghar <freghar@dummy.tld>