insta_zeus: add ACRE Spectator mode/keybind for Curator
ACRE2 has a "spectator" mode that disables normal direct speech and uses
a direction-less direct voice channel for "dead players". This is
normally activated via

  [true / false] call acre_api_fnc_setSpectator;

http://acre.idi-systems.com/api/api_general.html#acre_api_fnc_setSpectator

This mode also instructs ACRE to position direct voice of others in
relation to the camera, not the eyePos of the player unit, essentially
making the curator camera the listening point for direct chat.

In order to limit the "dead players" channel interaction with the GM, we
mute the channel for the GM by default. Anything said (using the direct
chat button) by the GM will unfortunately be heard in the spectator
channel, but at least anything said by dead players won't be heard by
the GM.

Due to how the ACRE spectator logic works (it's really meant only for
dead players), some limitations apply, such as the inability to transmit
on radios when in this mode. A better solution may be possible soon-ish:
https://github.com/IDI-Systems/acre2/issues/376

The logic also recognizes curator UI close/open and restores normal
behavior (speaking through player unit) when curator is closed, setting
the chosen mode back if the UI is re-opened (last setting is remembered).

All this functionality is disabled by default, but can be enabled by
a new CBA keybind under "Insta Zeus". This obviously works only when the
curator UI is open.

Signed-off-by: freghar <freghar@dummy.tld>