ares_extras: add Task Force functionality
This allows creation of "groups of groups" that share target knowledge
(enemy positions), with some delay. The primary use is direct fire
support (snipers, IFVs, static MGs, etc.), but it can be used for
simulated AI platoon-level comms or indirect fire support (with some
script that fires on known targets).

The effect is similar to grouping all units into a single group, but
without the movement restrictions (all units of a group want to stay
in formation).

To make a new TF from several groups, place the module on a ground,
select all relevant groups, hit return.

Selecting group(s) that already have a TF assigned will unassign them
from their original TFs, forming the new TF.

As such, to remove a group from a TF, place the module on the group
or place it on a ground and select only the one group.

Signed-off-by: freghar <freghar@dummy.tld>