silent_ops: squish west/east/indep operatives into one
You can still choose the side for the operative by joining him into
a group of appropriate side.

In Eden, this can be done by
  1) placing one or more SilentOperative units
  2) placing a temporary unit of the desired side
  3) grouping all SilentOperative units to that unit
  4) deleting the temporary unit
This leaves you with just the SilentOperative units in an appropriate
group, tied to a specific side.

The main advantage is that the units still have side=3 in the engine,
meaning they can wear any uniform which have side=3 in allowedSides{},
which are essentially all vanilla and modded uniforms.
(With added scripting, this enables undercover ops.)

Signed-off-by: freghar <freghar@dummy.tld>