execute_code: introduce Execute Code (Eden) inhibition logic
The main use case for Execute Code (Eden) is to pre-set some mission
defaults (settings), run some code once a custom composition is placed,
or similar.

The unexpected result, if the author doesn't delete the module from
within its code, is that the module code will be run in cases like
mission load (in 3DEN) or when editing/confirming the code, overwriting
any changes a mission maker might have done after placing the module!
The solution is to simply run the code once and delete the module.

However deleting the module prevents the author from making changes to
the code and saving a new version in a custom composition, as the module
literally disappears right after it's placed.

The solution here is an "inhibitor" module that, when present in
a mission, prevents any Execute Code (Eden) module from running its code
and from deleting itself when done.
After any changes to the Execute Code (Eden) module are done and saved
in appropriate custom compositions, the inhibitor module can be removed.
It should not be present in any saved missions as it serves no other
purpose than to temporarily stop the code execution / module deletion.

The inhibitor module has no effect on the runtime "Execute Code" module.

Signed-off-by: freghar <freghar@dummy.tld>