editor_extensions: remove ancient Arsenal module
This was the first module ever created in this mod and was probably never
properly used given that any official BI documentation points to a BI
function call and virtually nobody needs to provide a different interact
menu name for it.

Given that there's no ACE Arsenal, this module is doubly obsolete.

Signed-off-by: freghar <freghar@dummy.tld>