infopanel_control: allow disabling some features
"Info Panels" are the left/right mini boxes that typically show
(vanilla) GPS, vehicle radar, etc.

This allows disabling some of their components, so that, even if
a player has ie. ItemGPS and/or a vehicle has GPS enabled for the
driver, a GPS panel will not show.

Originally intended for simulating pre-GPS era scenarios.

Signed-off-by: freghar <freghar@dummy.tld>