editor_extensions/onelife: remove, is still buggy
The idea behind this was neat, but there were too many issues,
ie. weapon having separate physics and rolling away from the corpse
(and not being captured by getUnitLoadout) .. or the spectator not
always firing up correctly on JIP due to UI race conditions.
Or the 'pardon' functionality not working reliably due to buggy
teleports right after JIP.

So just remove it instead of dealing with all of this.

Use 'Kill player on JIP' for one-life missions using disabledAI=1
or 'Keep Playable AI' for one-life disabledAI=0 ones.

Signed-off-by: freghar <freghar@dummy.tld>