ee/post_process: remove, was too complex, buggy and useless
While really cool, it (even with defaults) spammed RPT with
"Wrong color format", even when it worked. Due to how post process
effects are stacked, this wasn't exactly easy to diagnose or debug.

The stacking and runtime control and 3DEN preview also added quite
a lot of complexity.

Couple all of that with how pretty useless the module is (only
for some occasional possible novelty/fanciness) and it makes sense
to just remove it.

Signed-off-by: freghar <freghar@dummy.tld>