ares_extras: remove 'Allow crew in immobile'
It was an "enable only" feature and since it's enabled by default
via ai_dynamic_skill, it doesn't make much sense to have it.

I could rework it to be an on/off, but it's such an obscure feature
that a global setting (ai_dynamic_skill CBA setting) is enough.

Signed-off-by: freghar <freghar@dummy.tld>