ee/teleportOnJIP: use damage disablement instead of simulation
The vanilla teleporting logic seems to wait for simulation to be
enabled so we cannot re-use it to prevent player unit damage.

This may lead to awkward situations of a player falling in air
when JIPed into a nonexistent plane/helicopter, so reset velocity
before teleport (to avoid glitching through objects @ dest).

Signed-off-by: freghar <freghar@dummy.tld>